Fixed the Ad Banner / View disappearing problem
When Ad Banner becomes full screen, the VC’s viewDidDisappear() is
invoked, and we free the GameView and the AdBanner view, thereby
calling for death.  Need to figure out how to free the VC’s memory when
we exit out of the game view controller.  Looks like there is a
circular cycle which is why the VC is being retained, perhaps by the
GameView due to the NSTimer…